requirements.txt
Django==1.8.1
django-treebeard==3.0
hash-ring==1.3.1
mongoengine==0.9.0
MySQL-python==1.2.5
Pillow==2.8.2
pymogile==2.0.2
pymongo==2.8
pysolr==3.3.1
pytz==2015.4
requests==2.7.0
solrpy==0.9.7
uWSGI==2.0.10
wheel==0.24.0
pyflakes==0.8.1
pep8==1.5.7
flake8==2.4.1
Fabric==1.10.2